build: defer making recovery patch
MAKE_RECOVERY_PATCH, or make_recovery_patch.py, calls GetBootableImage
to read recovery and boot images. It is supposed to write the SHA1
hash of recovery.img and boot.img to vendor/bin/install-recovery.sh
so that the integrity of the partition can be checked before applying
the patch to the recovery partition.

The problem of GetBootableImage is that it tries to make new boot
or recovery image if it cannot find one in target_files/BOOTABLE_IMAGES
or target_files/IMAGES. This means the hash in install-recovery.sh
is not the actual hash of the boot.img shipped in the final zip,
and applypatch will fail because of that.

Defer the process of making a recovery patch until the boot and
recovery images are ready in target_files. Since the patch may live
in vendor/ or system/vendor, build the patch right before building
system and vendor images.

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Change-Id: I2dace6e4841e0125b9f2713ed98cc57de9b38b4f
Signed-off-by: spkal01 <kalligeross@gmail.com>